Fixed a memory leak
The XML parser callback combined with the class level reference to the XML parser create a reference loop which can not be resolved by the PHP garbage collector. Therefore to prevent memory leaks either the class level reference should be cleaned (or not exist at all), or the parser callbacks should be unset (NULL) during cleanup. Just removing the class level reference, which as far as I can tell is not used anywhere, seems to be the easiest way.