Conditionally close a path when adding stroke
Fill/stroke operations are a "finalizing" instruction for paths in the PDF spec. As part of the stroke operation the path can be left open or closed. This was not accounted for in the Surface interface.

Additionally, CPDF was also always closing the path on a fillStroke operation. PDFLib does support both options, but only the open-path method was used.

This change updates the Surface interface with an optional argument for the stroke and fillStroke methods that indicates whether or not to close the path. The default is to leave the path open.